channels/line/pyproject.toml
[project]
name = "letsgo-channel-line"
version = "0.1.0"
description = "LINE channel adapter for LetsGo gateway"
requires-python = ">=3.11"
dependencies = [
    "letsgo-gateway",
]

[project.optional-dependencies]
sdk = ["line-bot-sdk>=3.0"]

[project.entry-points."letsgo.channels"]
line = "letsgo_channel_line:LINEChannel"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["letsgo_channel_line"]
